+++ STMSerial/Core/Inc/Tile.h
@@ -13,21 +13,21 @@
 class Tile : public CollidableObject {
 public:
 	Tile();
-	Tile(uint8_t posX, uint8_t posY, uint8_t tileType, bool isCollidable, bool isSpawnpoint);
+	Tile(uint16_t posX, uint16_t posY, uint8_t tileType, bool isCollidable, bool isSpawnpoint);
 	virtual ~Tile();
 	void setTileType(uint8_t type);
 	void setIsCollidable (bool isCollidable);
 	bool isCollidable() override;
-	uint8_t getPosX() override;
-	uint8_t getPosY() override;
+	uint16_t getPosX() override;
+	uint16_t getPosY() override;
 	pointVector getHalfSize() override;
 	uint8_t getTileType();
 	bool getIsCollidable();
 	bool getIsSpawnpoint();
 
 private:
-	uint8_t posX = 0;
-	uint8_t posY = 0;
+	uint16_t posX = 0;
+	uint16_t posY = 0;
 
 	uint8_t tileType = 0;
